Moved sorting code into the lib. There is now a uncommitted main file.
