+++ algorithm-practice/algor_data/sorting_test.py
@@ -1,4 +1,4 @@
-from .algor_data import data
+from . import data
 from .sorting import bubble_sort, insertion_sort, merge_sort
 
 import unittest
